requirements.txt
openai
requests
python-dotenv
smolagents
